+++ ErsatzTV.Core/FFmpeg/FFmpegLibraryProcessService.cs
@@ -404,13 +404,12 @@ public class FFmpegLibraryProcessService : IFFmpegProcessService
             graphicsElementContexts.AddRange(watermarks.Map(wm => new WatermarkElementContext(wm)));
         }
 
-        HardwareAccelerationMode decodeHwAccel = GetHardwareAccelerationMode(playbackSettings, fillerKind);
-        HardwareAccelerationMode encodeHwAccel = GetHardwareAccelerationMode(playbackSettings, FillerKind.None);
+        HardwareAccelerationMode hwAccel = GetHardwareAccelerationMode(playbackSettings);
 
         string videoFormat = GetVideoFormat(playbackSettings);
         Option<string> maybeVideoProfile = GetVideoProfile(videoFormat, channel.FFmpegProfile.VideoProfile);
         Option<string> maybeVideoPreset = GetVideoPreset(
-            encodeHwAccel,
+            hwAccel,
             videoFormat,
             channel.FFmpegProfile.VideoPreset,
             FFmpegLibraryHelper.MapBitDepth(channel.FFmpegProfile.BitDepth));
@@ -500,7 +499,7 @@ public class FFmpegLibraryProcessService : IFFmpegProcessService
 
         var desiredState = new FrameState(
             playbackSettings.RealtimeOutput,
-            fillerKind == FillerKind.Fallback,
+            InfiniteLoop: false,
             videoFormat,
             maybeVideoProfile,
             maybeVideoPreset,
@@ -549,10 +548,10 @@ public class FFmpegLibraryProcessService : IFFmpegProcessService
 
         var ffmpegState = new FFmpegState(
             saveReports,
-            decodeHwAccel,
-            encodeHwAccel,
-            VaapiDriverName(encodeHwAccel, vaapiDriver),
-            VaapiDeviceName(encodeHwAccel, vaapiDevice),
+            hwAccel,
+            hwAccel,
+            VaapiDriverName(hwAccel, vaapiDriver),
+            VaapiDeviceName(hwAccel, vaapiDevice),
             playbackSettings.StreamSeek,
             finish - now,
             channel.StreamingMode != StreamingMode.HttpLiveStreamingDirect,
@@ -577,16 +576,16 @@ public class FFmpegLibraryProcessService : IFFmpegProcessService
         _logger.LogDebug("FFmpeg desired state {FrameState}", desiredState);
 
         IPipelineBuilder pipelineBuilder = await _pipelineBuilderFactory.GetBuilder(
-            encodeHwAccel,
+            hwAccel,
             videoInputFile,
             audioInputFile,
             watermarkInputFile,
             subtitleInputFile,
             Option<ConcatInputFile>.None,
             graphicsEngineInput,
-            VaapiDisplayName(encodeHwAccel, vaapiDisplay),
-            VaapiDriverName(encodeHwAccel, vaapiDriver),
-            VaapiDeviceName(encodeHwAccel, vaapiDevice),
+            VaapiDisplayName(hwAccel, vaapiDisplay),
+            VaapiDriverName(hwAccel, vaapiDriver),
+            VaapiDeviceName(hwAccel, vaapiDevice),
             await customReportsFolder.IfNoneAsync(FileSystemLayout.FFmpegReportsFolder),
             FileSystemLayout.FontsCacheFolder,
             ffmpegPath);
@@ -701,7 +700,7 @@ public class FFmpegLibraryProcessService : IFFmpegProcessService
 
         var desiredState = new FrameState(
             playbackSettings.RealtimeOutput,
-            false,
+            InfiniteLoop: false,
             videoFormat,
             GetVideoProfile(videoFormat, channel.FFmpegProfile.VideoProfile),
             VideoPreset.Unset,
@@ -790,7 +789,7 @@ public class FFmpegLibraryProcessService : IFFmpegProcessService
         var videoInputFile = new VideoInputFile(videoPath, new List<VideoStream> { ffmpegVideoStream });
 
         // TODO: ignore accel if this already failed once
-        HardwareAccelerationMode hwAccel = GetHardwareAccelerationMode(playbackSettings, FillerKind.None);
+        HardwareAccelerationMode hwAccel = GetHardwareAccelerationMode(playbackSettings);
         _logger.LogDebug("HW accel mode: {HwAccel}", hwAccel);
 
         var ffmpegState = new FFmpegState(
@@ -1188,12 +1187,10 @@ public class FFmpegLibraryProcessService : IFFmpegProcessService
             .ForAccelAndFormat(hardwareAccelerationMode, videoFormat, bitDepth)
             .Find(p => string.Equals(p, videoPreset, StringComparison.OrdinalIgnoreCase));
 
-    private static HardwareAccelerationMode GetHardwareAccelerationMode(
-        FFmpegPlaybackSettings playbackSettings,
-        FillerKind fillerKind) =>
+    private static HardwareAccelerationMode GetHardwareAccelerationMode(FFmpegPlaybackSettings playbackSettings) =>
         playbackSettings.HardwareAcceleration switch
         {
-            _ when fillerKind == FillerKind.Fallback => HardwareAccelerationMode.None,
+            //_ when fillerKind == FillerKind.Fallback => HardwareAccelerationMode.None,
             HardwareAccelerationKind.Nvenc => HardwareAccelerationMode.Nvenc,
             HardwareAccelerationKind.Qsv => HardwareAccelerationMode.Qsv,
             HardwareAccelerationKind.Vaapi => HardwareAccelerationMode.Vaapi,